pyproject.toml
[project]
name = "wavedbg"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pyside6>=6.9.0",
    "pyvcd>=0.4.1",
]
